Removing linting issue
