+++ frontend/src/context/auth-context.ts
@@ -1,0 +1,16 @@
+import { createContext } from 'react';
+
+export type AuthUser = {
+  id: string;
+  email: string;
+  role: string;
+};
+
+export type AuthContextValue = {
+  token: string | null;
+  user: AuthUser | null;
+  login: (token: string, user: AuthUser) => void;
+  logout: () => void;
+};
+
+export const AuthContext = createContext<AuthContextValue | null>(null);
